Document, clean up code, make message a constant
Summary:
This commit doesn't introduce any new functionality, but makes our code nicer to work with.
It documents each function, moves the message that is sent to a constant String, and in general
cleans up our code a bit.

Test Plan:
I don't have access to an Arduino, so I can't test.

Reviewers: kg4sgp

CC: kd8lcv

Differential Revision: 12